+++ tim/matlab/matlab_renyi_entropy_lps_t.cpp
@@ -52,9 +52,9 @@ namespace
 			kNearestSuggestion,
 			range(filter.begin(), filter.end()));
 
-		const integer nans = std::max(estimate->t(), 0);
-		const integer skip = std::max(-estimate->t(), 0); 
-		const integer samples = std::max(nans + estimate->samples() - skip, 0);
+		const integer nans = std::max(estimate->t(), (integer)0);
+		const integer skip = std::max(-estimate->t(), (integer)0); 
+		const integer samples = std::max(nans + estimate->samples() - skip, (integer)0);
 
 		Array<real> result = createArray<real>(samples, 1, 
 			outputSet[Estimate]);
